scripts_python/requirements.txt
cassandra-driver=3.29.2